Time Series output MOHID Land changed
git-svn-id: https://mohid.svn.codeplex.com/svn@42749 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ Software/MOHIDLand/ModulePorousMediaProperties.F90
@@ -316,10 +316,10 @@ Module ModulePorousMediaProperties
 
     type T_Property
         type (T_PropertyID)                     :: ID
-        real(8), dimension(:,:,:), pointer      :: Concentration            => null()
-        real(8), dimension(:,:,:), pointer      :: ConcentrationOld         => null()
-        real(8), dimension(:,:), pointer        :: ConcentrationOnInfColumn      => null()
-        real(8), dimension(:,:), pointer        :: ConcentrationDN               => null()
+        real, dimension(:,:,:), pointer         :: Concentration            => null()
+        real, dimension(:,:,:), pointer         :: ConcentrationOld         => null()
+        real, dimension(:,:), pointer           :: ConcentrationOnInfColumn      => null()
+        real, dimension(:,:), pointer           :: ConcentrationDN               => null()
         real, pointer, dimension(:,:,:)         :: Mass_Created
         real(8),    pointer, dimension(:,:,:)   :: ViscosityU
         real(8),    pointer, dimension(:,:,:)   :: ViscosityV
